requirements.txt
streamlit>=1.45.0
pymupdf>=1.25.0
sentence-transformers>=3.4.0
faiss-cpu>=1.10.0
groq>=0.23.0
python-dotenv>=1.1.0
numpy>=2.0.0
pandas>=2.2.0
tf-keras
psutil